+++ c/reducercommutativity/rangesum_false-unreach-call.c
@@ -20,7 +20,7 @@ int N;
 int rangesum (int x[N])
 {
   int i;
-  int ret;
+  long long ret;
   ret = 0;
   int cnt = 0;
   for (i = 0; i < N; i++) {
